# lib/units/api/swagger/api_v1.yaml
openapi: 3.0.1
info:
  title: DeviceHub
  description: Control and manages real Smartphone devices from browser and restful
    apis
  contact:
    name: Devicehub Team
    url: https://github.com/VKCOM/devicehub
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: 1.3.10
servers:
- url: /api/v1
...
paths:
  /groups:
    get:
      tags:
      - groups
      summary: Gets groups
      description: Returns the groups to which you belong
      operationId: getGroups
      parameters:
      - name: fields
        in: query
        description: Comma-seperated list of fields; only listed fields will be returned
          in response
        schema:
          type: string
      - name: owner
        in: query
        description: "Selects the groups for which you are the owner (true) or a simple\
          \ member (false); note that by not providing this parameter, it means all\
          \ groups to which you belong are selected"
        schema:
          type: boolean
      responses:
        "200":
          description: Groups information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupListResponse'
        default:
          description: |
            Unexpected Error:
              * 401: Unauthorized => bad credentials
              * 500: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedErrorResponse'
    post:
      tags:
      - groups
      summary: Creates a group
      description: Creates a group with you as owner
      operationId: createGroup
      requestBody:
        description: Group properties; at least one property is required
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupPayload'
          application/octet-stream:
            schema:
              $ref: '#/components/schemas/GroupPayload'
        required: true
      responses:
        "201":
          description: Group information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        default:
          description: |
            Unexpected Error:
              * 400: Bad Request => invalid format or semantic of properties
              * 401: Unauthorized => bad credentials
              * 500: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedErrorResponse'
      x-codegen-request-body-name: group
    delete:
      tags:
      - groups
      summary: Removes groups
      description: Removes the groups owned by you
      operationId: deleteGroups
      requestBody:
        description: Groups to remove as a comma-separated list of group identifiers;
          note that by not providing this parameter it means all groups owned by you
          are removed
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupsPayload'
          application/octet-stream:
            schema:
              $ref: '#/components/schemas/GroupsPayload'
        required: false
      responses:
        "200":
          description: Groups removing is OK (or no groups to remove)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponse'
        default:
          description: |
            Unexpected Error:
              * 401: Unauthorized => bad credentials
              * 403: Forbidden => a device is currently booked or unremovable built-in group
              * 404: Not Found => unknown groups
              * 500: Internal Server Error
              * 503: Service Unavailable => server too busy or a lock on a resource is pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedErrorResponse'
      x-codegen-request-body-name: groups
    x-swagger-router-controller: groups
...
  /groups/{id}/devices:
    get:
      tags:
      - groups
      summary: Gets the devices of a group
      description: Returns the devices of the group to which you belong
      operationId: getGroupDevices
      parameters:
      - name: id
        in: path
        description: Group identifier
        required: true
        schema:
          type: string
      - name: bookable
        in: query
        description: "Selects devices which could be potentially booked by that transient\
          \ group (true => irrelevant for an origin group!), or selects all devices\
          \ of the group (false); note that by not providing this parameter all devices\
          \ of the group are selected"
        schema:
          type: boolean
          default: false
      - name: fields
        in: query
        description: Comma-separated list of device fields; only listed fields will
          be returned in response
        schema:
          type: string
      responses:
        "200":
          description: Group devices information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceListResponse'
        default:
          description: |
            Unexpected Error:
              * 400: Bad request => group is not transient
              * 401: Unauthorized => bad credentials
              * 404: Not Found => unknown group
              * 500: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedErrorResponse'
    put:
      tags:
      - groups
      summary: Adds devices into a transient group
      description: Adds devices into a transient group owned by you
      operationId: addGroupDevices
      parameters:
      - name: id
        in: path
        description: Group identifier
        required: true
        schema:
          type: string
      requestBody:
        description: Devices to add as a comma-separated list of serials; note that
          by not providing this parameter it means all devices which could be potentially
          booked by that transient group are added into the latter
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevicesPayload'
          application/octet-stream:
            schema:
              $ref: '#/components/schemas/DevicesPayload'
        required: false
      responses:
        "200":
          description: Group information (an empty group is returned if no change
            is made)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        "409":
          description: Conflicts information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictsResponse'
        default:
          description: |
            Unexpected Error:
              * 400: Bad Request => group is not transient
              * 401: Unauthorized => bad credentials
              * 403: Forbidden => quota is reached
              * 404: Not Found => unknown group or devices
              * 500: Internal Server Error
              * 503: Service Unavailable => server too busy or a lock on a resource is pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedErrorResponse'
      x-codegen-request-body-name: devices
    delete:
      tags:
      - groups
      summary: Removes devices from a transient group
      description: Removes devices from a transient group owned by you
      operationId: removeGroupDevices
      parameters:
      - name: id
        in: path
        description: Group identifier
        required: true
        schema:
          type: string
      requestBody:
        description: Devices to remove as a comma-separated list of serials; note
          that by not providing this parameter it means all devices of the group are
          removed
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevicesPayload'
          application/octet-stream:
            schema:
              $ref: '#/components/schemas/DevicesPayload'
        required: false
      responses:
        "200":
          description: Group information (an empty group is returned if no change
            is made)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        default:
          description: |
            Unexpected Error:
              * 400: Bad Request => group is not transient
              * 401: Unauthorized => bad credentials
              * 404: Not Found => unknown group or devices
              * 500: Internal Server Error
              * 503: Service Unavailable => server too busy or a lock on a resource is pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedErrorResponse'
      x-codegen-request-body-name: devices
    x-swagger-router-controller: groups
...
  /user/devices:
    get:
      tags:
      - user
      summary: User Devices
      description: The User Devices endpoint returns device list owner by current
        authorized user
      operationId: getUserDevices
      parameters:
      - name: fields
        in: query
        description: Fields query parameter takes a comma seperated list of fields.
          Only listed field will be return in response
        schema:
          type: string
      responses:
        "200":
          description: Current User Devices List
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceListResponse'
        default:
          description: |
            Unexpected Error:
              * 401: Unauthorized => bad credentials
              * 500: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedErrorResponse'
    post:
      tags:
      - user
      summary: Add a device to a user
      description: The User Devices endpoint will request stf server for a new device.
      operationId: addUserDevice
      requestBody:
        description: Device to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUserDevicePayload'
          application/octet-stream:
            schema:
              $ref: '#/components/schemas/AddUserDevicePayload'
        required: true
      responses:
        "200":
          description: Add User Device Status
          content: {}
        default:
          description: |
            Unexpected Error:
              * 401: Unauthorized => bad credentials
              * 403: Forbidden => device is already controlled or is not available
              * 404: Not Found => unknown device
              * 500: Internal Server Error
              * 504: Gateway Time-out => device is not responding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedErrorResponse'
      x-codegen-request-body-name: device
    x-swagger-router-controller: user
...
  /user/devices/{serial}/remoteConnect:
    post:
      tags:
      - user
      summary: Remote Connect
      description: The device connect endpoint will request stf server to connect
        remotely
      operationId: remoteConnectUserDeviceBySerial
      parameters:
      - name: serial
        in: path
        description: Device Serial
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Remote Connect User Device Request Status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoteConnectUserDeviceResponse'
        default:
          description: |
            Unexpected Error:
              * 401: Unauthorized => bad credentials
              * 403: Forbidden => device is not controlled by the user
              * 404: Not Found => unknown device
              * 500: Internal Server Error
              * 504: Gateway Time-out => device is not responding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedErrorResponse'
    delete:
      tags:
      - user
      summary: Remote Disconnect
      description: The device connect endpoint will request stf server to disconnect
        remotely
      operationId: remoteDisconnectUserDeviceBySerial
      parameters:
      - name: serial
        in: path
        description: Device Serial
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Remote Disсonnect User Device Request Status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponse'
        default:
          description: |
            Unexpected Error:
              * 401: Unauthorized => bad credentials
              * 403: Forbidden => device is not controlled by the user
              * 404: Not Found => unknown device
              * 500: Internal Server Error
              * 504: Gateway Time-out => device is not responding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedErrorResponse'
    x-swagger-router-controller: user
...
  /autotests:
    get:
      tags:
      - autotests
      summary: Get devices for autotests run
      description: Create group and return serials of captured devices
      operationId: captureDevices
      parameters:
      - name: timeout
        in: query
        description: timeout for group in seconds
        required: true
        schema:
          minimum: 600
          type: integer
      - name: amount
        in: query
        description: Device amount needed for autotests run
        required: true
        schema:
          minimum: 1
          type: integer
      - name: need_amount
        in: query
        description: "need only specified amount, not less"
        schema:
          type: boolean
      - name: abi
        in: query
        description: device abi
        schema:
          type: string
      - name: sdk
        in: query
        description: device sdk
        schema:
          type: string
      - name: model
        in: query
        description: device model
        schema:
          type: string
      - name: type
        in: query
        description: device type
        schema:
          type: string
      - name: version
        in: query
        description: device os version
        schema:
          type: string
      - name: run
        in: query
        description: run identificator
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Group created for run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        default:
          description: |
            Unexpected Error:
              * 401: Unauthorized => bad credentials
              * 404: Not Found => unknown data
              * 500: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedErrorResponse'
    delete:
      tags:
      - autotests
      summary: delete autotests run data
      description: delete group and data created for autotests
      operationId: freeDevices
      parameters:
      - name: group
        in: query
        description: Device amount needed for autotests run
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Group created for run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponse'
        default:
          description: |
            Unexpected Error:
              * 401: Unauthorized => bad credentials
              * 404: Not Found => unknown data
              * 500: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedErrorResponse'
    x-swagger-router-controller: autotests
  /autotests/{id}/addDevices/:
      get:
          tags:
              - autotests
          summary: Add devices for autotests group
          description: Filter and add devices to autotests group
          operationId: addDevices
          parameters:
              -   name: id
                  in: path
                  description: Group identifier
                  required: true
                  schema:
                      type: string
              -   name: amount
                  in: query
                  description: Device amount needed for autotests run
                  required: true
                  schema:
                      minimum: 1
                      type: integer
              -   name: need_amount
                  in: query
                  description: "need only specified amount, not less"
                  schema:
                      type: boolean
              -   name: abi
                  in: query
                  description: device abi
                  schema:
                      type: string
              -   name: sdk
                  in: query
                  description: device sdk
                  schema:
                      type: string
              -   name: model
                  in: query
                  description: device model
                  schema:
                      type: string
              -   name: type
                  in: query
                  description: device type
                  schema:
                      type: string
              -   name: version
                  in: query
                  description: device os version
                  schema:
                      type: string
          responses:
              "200":
                  description: Updated group with new devices
                  content:
                      application/json:
                          schema:
                              $ref: '#/components/schemas/GroupResponse'
              default:
                  description: |
                      Unexpected Error:
                        * 401: Unauthorized => bad credentials
                        * 404: Not Found => unknown data
                        * 500: Internal Server Error
                  content:
                      application/json:
                          schema:
                              $ref: '#/components/schemas/UnexpectedErrorResponse'
          x-codegen-request-body-name: addDevices
      x-swagger-router-controller: autotests
  /autotests/install/{serial}:
    post:
      tags:
      - autotests
      summary: install apk on device by serial
      description: Installing apk to device from url
      operationId: installOnDevice
      parameters:
      - name: serial
        in: path
        description: Device identifier (serial)
        required: true
        schema:
          type: string
      requestBody:
        description: List of flags for adb install
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdbInstallFlagsPayload'
          application/octet-stream:
            schema:
              $ref: '#/components/schemas/AdbInstallFlagsPayload'
        required: true
      responses:
        "200":
          description: Remote Connect User Device Request Status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoteConnectUserDeviceResponse'
        default:
          description: |
            Unexpected Error:
              * 401: Unauthorized => bad credentials
              * 403: Forbidden => device is not controlled by the user
              * 404: Not Found => unknown device
              * 500: Internal Server Error
              * 504: Gateway Time-out => device is not responding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedErrorResponse'
      x-codegen-request-body-name: installProperties
    x-swagger-router-controller: autotests
...
components:
  schemas:
    UnexpectedErrorResponse:
      required:
      - description
      - success
      type: object
      properties:
        success:
          type: boolean
          default: false
        description:
          type: string
...
    GroupResponse:
      required:
      - description
      - group
      - success
      type: object
      properties:
        success:
          type: boolean
        description:
          type: string
        group:
          type: object
          properties: {}
          description: A null value means the group is unchanged
...
    GroupPayload:
      type: object
      properties:
        name:
          pattern: "^[0-9a-zA-Z-_./: ]{1,50}$"
          type: string
          description: Group Name; default value => generated at runtime
        startTime:
          type: string
          description: "Group starting time (in UTC, conforming to RFC 3339 section\
            \ 5.6); default value => group creation time"
          format: date-time
        stopTime:
          type: string
          description: "Group expiration time (in UTC, conforming to RFC 3339 section\
            \ 5.6); default value => startTime + 1 hour"
          format: date-time
        class:
          type: string
          description: "Group class; privileged value => debug, bookable, standard"
          enum:
          - once
          - bookable
          - hourly
          - daily
          - weekly
          - monthly
          - quaterly
          - halfyearly
          - yearly
          - debug
          - standard
        repetitions:
          minimum: 0
          type: integer
          description: Group repetitions; default value => 0
        state:
          type: string
          description: Group state; default value => pending or ready for bookable/standard
            classes
          enum:
          - pending
          - ready
      description: Payload object for creating/updating a group
    GroupsPayload:
      type: object
      properties:
        ids:
          type: string
          description: Comma-separated list of identifiers
      description: Payload object for adding/removing groups
...